separate themes and theme colors
